prepare theme file for build
