+++ lib/hooks/useTheme.ts
@@ -1,9 +1,9 @@
 import { useTheme as useEmotionContextTheme } from "@emotion/react";
 import theme from "../assets/theme";
-import { RGTheme } from "../types/Theme";
 import merge from "lodash.merge";
 import { useReactGridStore } from "../utils/reactGridStore";
 import { useReactGridId } from "../components/ReactGridIdProvider";
+import { RGTheme } from "../types/RGTheme";
 
 export const useTheme = (): RGTheme => {
   const id = useReactGridId();
